requirements.txt
numpy
pillow
torch
transformers
opencv-contrib-python
scikit-learn
matplotlib
pyyaml
pypylon
huggingface-hub
